+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Added
 
+- the Turtle encoder can now produce partial Turtle documents with the `:only` option
+  and any combination of the following values: `:triples`, `:directives`, `:base`, `:prefixes`  
 - the style of the Turtle directives produced by the Turtle encoder can be
   switched to SPARQL style with the option `:directive_style` and the value `:sparql`